@@ -5,7 +5,6 @@ import logging
 import contracts
 from zuper_commons.types import ZValueError
 from zuper_commons.text import remove_escapes
-from zuper_typing import debug_print
 
 from pdm4ar.available_exercises import available_exercises
 from pdm4ar.exercises_def.structures import Exercise, ExerciseEvaluator, out_dir
@@ -37,7 +36,7 @@ def run_exercise(exercise: str, evaluation_mode=False):
 
     report_file = os.path.join(out_main, "index.html")
     report.to_html(report_file)
-    result_str = remove_escapes(debug_print(perf))
+    result_str = remove_escapes(str(perf))
     logger.info(f"\n<<<<<\n{result_str}\n>>>>>")
 
 
